requirements.txt
discord.py==1.7.3
loguru==0.5.3
python-dotenv==0.18.0
psutil==5.8.0
distro==1.5.0
pyshorteners==1.0.1
beautifulsoup4==4.10.0
pytz==2021.1